# Reciprocal-based 2by1 division procedure