This is based on [Improved division by invariant integers](https://gmplib.org/~tege/division-paper.pdf) and ported from https://github.com/chfast/intx.

The computation of reciprocal is single `DIV` x86_64 instruction, but there exist faster implementation (sneak peek in #39) but I left it for other time.

```
name                     old time/op  new time/op  delta
_Div/large/big-6          246ns ± 0%   245ns ± 0%   -0.49%  (p=0.016 n=4+5)
_Div/large/uint256-6     99.2ns ± 0%  83.3ns ± 0%  -16.07%  (p=0.008 n=5+5)
_Div/small/big-6         68.8ns ± 0%  68.8ns ± 0%     ~     (p=0.500 n=5+5)
_Div/small/uint256-6     13.1ns ± 0%  13.0ns ± 0%   -0.76%  (p=0.008 n=5+5)
_MulMod/large/big-6       388ns ± 0%   388ns ± 0%     ~     (p=0.167 n=5+5)
_MulMod/large/uint256-6   204ns ± 0%   153ns ± 0%  -25.00%  (p=0.008 n=5+5)
_MulMod/small/big-6      94.1ns ± 1%  93.9ns ± 0%     ~     (p=0.397 n=5+5)
_MulMod/small/uint256-6  40.0ns ± 0%  40.0ns ± 0%     ~     (p=0.444 n=5+5)
MulMod/mod64-6            186ns ± 0%   101ns ± 0%  -45.70%  (p=0.008 n=5+5)
MulMod/mod128-6           266ns ± 0%   176ns ± 0%  -33.68%  (p=0.000 n=4+5)
MulMod/mod192-6           245ns ± 0%   173ns ± 0%     ~     (p=0.079 n=4+5)
MulMod/mod256-6           223ns ± 0%   169ns ± 0%  -24.22%  (p=0.000 n=5+4)
_Mod/large/big-6          137ns ± 0%   137ns ± 0%     ~     (all equal)
_Mod/large/uint256-6     50.2ns ± 0%  68.2ns ± 0%  +35.94%  (p=0.016 n=4+5)
_Mod/small/big-6         40.8ns ± 0%  40.9ns ± 0%     ~     (p=0.325 n=5+5)
_Mod/small/uint256-6     14.0ns ± 0%  14.0ns ± 0%     ~     (all equal)
_SDiv/large/big-6         402ns ± 0%   401ns ± 0%   -0.25%  (p=0.016 n=4+5)
_SDiv/large/uint256-6     105ns ± 0%    89ns ± 0%  -14.95%  (p=0.000 n=5+4)
_AddMod/large/big-6       212ns ± 0%   212ns ± 0%     ~     (all equal)
_AddMod/large/uint256-6  39.5ns ± 0%  39.9ns ± 1%   +0.96%  (p=0.024 n=5+5)
_AddMod/small/big-6      65.4ns ± 0%  65.3ns ± 1%     ~     (p=0.238 n=5+5)
_AddMod/small/uint256-6  17.2ns ± 0%  17.4ns ± 0%   +1.16%  (p=0.008 n=5+5)
```