@holiman 2020-03-18
Cool. But this stood out: 
```
_Mod/large/uint256-6     50.2ns ± 0%  68.2ns ± 0%  +35.94%  (p=0.016 n=4+5)
```
What happened there - any idea?