@chfast 2020-03-18
> Cool. But this stood out:
> 
> ```
> _Mod/large/uint256-6     50.2ns ± 0%  68.2ns ± 0%  +35.94%  (p=0.016 n=4+5)
> ```
> 
> What happened there - any idea?

I noticed it too late. Nevertheless, I can explain it. The numbers are suspicious anyway if you compare with `_Div/large/uint256-6     99.2ns ± 0%  83.3ns ± 0%  -16.07%` (exactly the same algo).

The benchmark is `0xfefefefefefefefe... mod 0xefefefefefefefef...`. The divisor is already normalized. The Knuth division requires only single step where it divides `<0, 0xfefefefefefefefe> / 0xefefefefefefefef` using single `DIV` instruction.

After the change we always compute the reciprocal, i.e. `<^d, 0xffffffffffffffff> / d` using the same single `DIV` instruction. Then any later division is computed by multiplication + additional adjustments.

So this is more work in this case for the new implementation. But there is more.

If you check the `DIV` instruction it is actually some kind of algorithm: 36 microinstructions, 32-96 cycles (Haswell), 35-88 cycles (Skylake). And when the numerator high part is zero (as when executing the benchmark in old implementation) it finishes quicker (verified empirically some time ago).

In summary, division is tricky because of many special cases in implementation. The same benchmarking it is tricky. That's why I prefer to generate multiple random samples and benchmark full samples set. Also: the complexity is `O(len(numerator) - len(divisor) + 1)` so 256 by 256 is not the worst case. And division by single word is implemented differently, division by two words _can_ be implemented differently.

I can rework benchmarking to do the same as in new `BenchmarkMulMod` using 1024 samples. This PR can be merged after.