@chfast 2020-03-19
More detailed Mod benchmarks in #42.